resize_image moved to BinaryImageField, minor changes
