+++ chatup/api/chat/serializers.py
@@ -21,8 +21,8 @@ class CustomBinaryImageField(BinaryImageField):
 
     def __init__(self, *args, **kwargs):
         image_type = self.context['request'].data['type']
-        image_size = models.Image.SIZES[image_type]
-        super().__init__(size=image_size, *args, **kwargs)
+        kwargs['size'] = models.Image.SIZES[image_type]
+        super().__init__(*args, **kwargs)
 
 
 class RoleSerializer(TranslatedModelSerializer):
@@ -136,8 +136,8 @@ class ImageSerializer(ImageCacheSerializer):
         if request.method in UPDATE_METHODS:
             self.fields['type'].read_only = True
         elif (
-                request.method in SAFE_METHODS
-                and (not user_role or user_role.sid not in {models.Role.SIDS.ADMIN, models.Role.SIDS.STREAMER})
+            request.method in SAFE_METHODS
+            and (not user_role or user_role.sid not in {models.Role.SIDS.ADMIN, models.Role.SIDS.STREAMER})
         ):
             self.fields.pop('users')
